requirements.txt
python-telegram-bot==20.7
yt-dlp
aiohttp
redis[asyncio]
python-dotenv
